requirements.txt
feedparser==6.0.11
beautifulsoup4==4.13.5
python-dotenv==1.1.1
langchain-openai==0.3.31
langchain-community==0.3.27
langchain-core==0.3.74
flask==3.0.0
sentence-transformers==5.1.0
langchain-huggingface==0.3.1
chromadb
